+++ libs/ng-bpmn/src/lib/utils/debounce.ts
@@ -1,0 +1,13 @@
+export function debounce(fn: (...args: any[]) => void, timeout = 500) {
+  let timer: number;
+
+  return function (...args: any[]) {
+    if (timer) {
+      clearTimeout(timer);
+    }
+
+    timer = setTimeout(() => {
+      fn(...args);
+    }, timeout);
+  };
+}